Add idapi mocking as custom command
