+++ cypress/integration/reset_password.spec.js
@@ -3,34 +3,22 @@
 const PageResetPassword = require('../support/pages/reset_password_page');
 const PageResetSent = require('../support/pages/reset_sent_page');
 
-const MOCKING_ENDPOINT = 'localhost:9000/mock';
-
-const getMockOptions = (status, body = {}) => ({
-  headers: {
-    'Content-Type': 'application/json',
-    "x-status": status,
-  },
-  method: 'POST',
-  body: JSON.stringify(body),
-  url: MOCKING_ENDPOINT, 
-});
-
 describe("Password reset flow", () => {
   const page = new PageResetPassword();
 
   before(() => {
-    cy.request(MOCKING_ENDPOINT + '/purge');
+    cy.idapiMockPurge();
   });
 
-  beforeEach(function() {
+  beforeEach(function () {
     cy.fixture('users').as('users');
     page.goto();
   });
-  
+
   context("Valid email already exits", () => {
     it("successfully submits the request", function () {
       const { email } = this.users.validEmail;
-      cy.request(getMockOptions(200)); 
+      cy.idapiMock(200);
       page.submitEmailAddress(email);
       cy.contains(PageResetSent.CONTENT.CONFIRMATION);
     });
@@ -39,10 +27,10 @@ describe("Password reset flow", () => {
   context("Email doesn't exist", () => {
     it('shows a message saying the email address does not exist', function () {
       const { email } = this.users.emailNotRegistered;
-      cy.request(getMockOptions(404, { 
-        status: 'error', 
-        errors: [{message: 'Not found'}]
-      })); 
+      cy.idapiMock(404, {
+        status: 'error',
+        errors: [{ message: 'Not found' }]
+      });
       page.submitEmailAddress(email);
       cy.contains(PageResetPassword.CONTENT.ERRORS.NO_ACCOUNT);
     });
@@ -50,10 +38,10 @@ describe("Password reset flow", () => {
 
   context("Email field is left blank", () => {
     it('displays a message saying an email address is needed', () => {
-      cy.request(getMockOptions(500, { 
-        status: 'error', 
-        errors: [{message: 'Required field missing'}]
-      })); 
+      cy.idapiMock(500, {
+        status: 'error',
+        errors: [{ message: 'Required field missing' }]
+      });
       page.clickResetPassword();
       cy.contains(PageResetPassword.CONTENT.ERRORS.NO_EMAIL);
     });
@@ -61,7 +49,7 @@ describe("Password reset flow", () => {
 
   context("General IDAPI failure", () => {
     it('displays a generic error message', () => {
-      cy.request(getMockOptions(500));
+      cy.idapiMock(500);
       page.clickResetPassword();
       cy.contains(PageResetPassword.CONTENT.ERRORS.GENERIC);
     });
